Update model and admin
Create a token generator which can be use to apply default value on token value in the admin panel.
Create a new test runner to run test without database.
Create test for the token generator and validate that it returns a string with the right length.
Update the token field for the form to be displayed a password field.